[update] adding common structure for all projects
