--- src/assets/css/main.css
@@ -1,179 +1,0 @@
-.semi-bold{
-    font-weight: 500;
-}
-
-.text-right,
-.text-right .webix_el_box{
-    text-align: right;
-}
-
-.text-center,
-.text-center .webix_el_box{
-    text-align: center;
-}
-
-.uppercase,
-.uppercase .webix_el_box{
-    text-transform: uppercase;
-}
-
-.gray,
-.gray .webix_el_box{
-    color: #818181;
-}
-
-.font-16,
-.font-16 .webix_el_box{
-    font-size: 16px;
-}
-
-.font-18,
-.font-18 .webix_el_box{
-    font-size: 18px;
-}
-
-.item-violet{
-    color: #6776e7;
-}
-
-.item-violet .list-item-icon{
-    background-color: #e9eefe;
-}
-
-.item-red,
-.item-red .webix_el_box{
-    color: #fd4b51;
-}
-
-.item-red .list-item-icon{
-    background-color: #ffdbdc;
-}
-
-.item-green{
-    color: #52c083;
-}
-
-.item-green .list-item-icon{
-    background-color: #e3faef;
-}
-
-.item-orange{
-    color: #faa57b;
-}
-
-.item-orange .list-item-icon{
-    background-color: #fff5e9;
-}
-
-.item-blue{
-    color: #5095e0;
-}
-
-.item-blue .list-item-icon{
-    background-color: #e1f0fe;
-}
-
-.label-secondary,
-.label-secondary > .webix_el_box{    
-    font-size: 11px;
-    padding-bottom: 0;
-}
-
-.color-item{
-    display: inline-block;
-    min-width: 60px;
-    padding: 5px 10px;
-
-
-    border-radius: 30px;
-
-    font-size: 12px;
-    font-weight: 500;
-    text-align: center;
-    color: #fff;
-}
-
-.color-red{
-    background-color: #e95567;
-}
-
-.color-yellow{
-    background-color: #ffcc30;
-}
-
-.color-blue{
-    background-color: #5095e0 ;
-}
-
-.color-green{
-    background-color: #40e48a;
-}
-
-.color-bright-blue{
-    background-color: #2b6ce4;
-}
-
-.color-gray{
-    background-color: #adb9c4;
-}
-
-.webix_el_icon .webix_disabled{
-    cursor: default;
-}
-
-.webix_view.webix_layout_custom {
-    background-color: #fff;
-}
-
-.template-main-panel {
-    padding-bottom: 10px;
-    padding-top: 3px;
-    font-weight: 500;
-}
-
-.custom-table .webix_column>div, 
-.custom-table .webix_table_cell {
-    padding: 10px 20px;
-    border-bottom: 1px solid transparent;
-
-    line-height: normal !important;
-}
-
-.custom-table .webix_hcell.table-gray-header {
-    color: #818181;
-    background-color: #f0f1f3;
-    font-size: 12px;
-    font-weight: 500;
-    padding: 0 20px;
-}
-
-.custom-table .webix_ss_header td.webix_last_row,
-.custom-table div.webix_ss_vscroll_header {
-    background-color: #f0f1f3;
-    border-bottom: 1px solid transparent;    
-    border-left: 1px solid transparent;
-}
-
-.custom-table-clean .webix_column>div, 
-.custom-table-clean .webix_table_cell, 
-.custom-table-clean .webix_hcell {
-    padding-left: 20px;
-    border-bottom: 1px solid transparent;
-}
-
-.custom-table-clean .webix_hcell.header-multiline {
-    white-space: normal;
-    line-height:15px !important;
-    padding-top:12px;
-}
-
-.custom-table-clean .webix_ss_header td.webix_last_row,
-.custom-table-clean div.webix_ss_vscroll_header {
-    background-color: #fff;
-    border-bottom: 1px solid transparent;    
-    border-left: 1px solid transparent;
-}
-.custom-table-clean .webix_cell {
-    white-space: normal;
-    padding-top:12px;
-}
