+++ src/index.html
@@ -4,11 +4,11 @@
 	<meta charset="UTF-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<meta http-equiv="X-UA-Compatible" content="ie=edge">
-	<title>Webpack</title>
+	<title>Personal tasks board</title>
 	<!-- Webix Library -->
 	<script type="text/javascript" src="//cdn.webix.com/pro/edge/webix.js"></script>
 	<link rel="stylesheet" type="text/css" href="//cdn.webix.com/pro/edge/webix.css">
-
+	<!-- Material Design icons -->
 	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/MaterialDesign-Webfont/4.9.95/css/materialdesignicons.min.css">
 </head>
 <body>
