+++ src/styles/_list.scss
@@ -76,11 +76,6 @@
     font-size: 14px;
 }
 
-.first-col{
-    display: flex;
-    padding-left:20px;
-}
-
 .list-icon-check i{
     line-height: 1.6;
     display: inline-block;
