+++ src/styles/_listActions.scss
@@ -1,0 +1,40 @@
+.item-violet{
+    color: #6776e7;
+}
+
+.item-violet .list-item-icon{
+    background-color: #e9eefe;
+}
+
+.item-red,
+.item-red .webix_el_box{
+    color: #fd4b51;
+}
+
+.item-red .list-item-icon{
+    background-color: #ffdbdc;
+}
+
+.item-green{
+    color: #52c083;
+}
+
+.item-green .list-item-icon{
+    background-color: #e3faef;
+}
+
+.item-orange{
+    color: #faa57b;
+}
+
+.item-orange .list-item-icon{
+    background-color: #fff5e9;
+}
+
+.item-blue{
+    color: #5095e0;
+}
+
+.item-blue .list-item-icon{
+    background-color: #e1f0fe;
+}
