+++ src/styles/_listBadges.scss
@@ -1,0 +1,37 @@
+.color-item{
+    display: inline-block;
+    min-width: 60px;
+    padding: 5px 10px;
+
+
+    border-radius: 30px;
+
+    font-size: 12px;
+    font-weight: 500;
+    text-align: center;
+    color: #fff;
+}
+
+.color-red{
+    background-color: #e95567;
+}
+
+.color-yellow{
+    background-color: #ffcc30;
+}
+
+.color-blue{
+    background-color: #5095e0 ;
+}
+
+.color-green{
+    background-color: #40e48a;
+}
+
+.color-bright-blue{
+    background-color: #2b6ce4;
+}
+
+.color-gray{
+    background-color: #adb9c4;
+}
