+++ src/styles/_tableAssignments.scss
@@ -1,0 +1,23 @@
+.custom-table-clean .webix_column>div, 
+.custom-table-clean .webix_table_cell, 
+.custom-table-clean .webix_hcell {
+    padding-left: 20px;
+    border-bottom: 1px solid transparent;
+}
+
+.custom-table-clean .webix_hcell.header-multiline {
+    white-space: normal;
+    line-height:15px !important;
+    padding-top:12px;
+}
+
+.custom-table-clean .webix_ss_header td.webix_last_row,
+.custom-table-clean div.webix_ss_vscroll_header {
+    background-color: #fff;
+    border-bottom: 1px solid transparent;    
+    border-left: 1px solid transparent;
+}
+.custom-table-clean .webix_cell {
+    white-space: normal;
+    padding-top:12px;
+}
