+++ src/styles/_tableCommon.scss
@@ -1,0 +1,21 @@
+.custom-table .webix_column>div, 
+.custom-table .webix_table_cell {
+    padding: 10px 20px;
+    border-bottom: 1px solid transparent;
+    line-height: normal !important;
+}
+
+.custom-table .webix_hcell.table-gray-header {
+    color: #818181;
+    background-color: #f0f1f3;
+    font-size: 12px;
+    font-weight: 500;
+    padding: 0 20px;
+}
+
+.custom-table .webix_ss_header td.webix_last_row,
+.custom-table div.webix_ss_vscroll_header {
+    background-color: #f0f1f3;
+    border-bottom: 1px solid transparent;    
+    border-left: 1px solid transparent;
+}
